requirements.txt
PyGithub==1.57
PyYAML==5.4.1